requirements-dev.txt
black
pytest==7.4.0
pytest-aiohttp==1.0.4
pytest-cov==4.1.0
-e .
